+++ src/idmodels/__init__.py
@@ -1,2 +1,11 @@
-__version__ = "1.2.0"
+from idmodels.config import (DataSource, Disease, GBQRModelConfig, GBQRRunConfig, PoolingStrategy, PowerTransform,
+                             SARIXFourierModelConfig, SARIXModelConfig, SARIXRunConfig)
+from idmodels.gbqr import GBQRModel
+from idmodels.sarix import SARIXFourierModel, SARIXModel
 
+
+__all__ = ['DataSource', 'Disease', 'GBQRModel', 'GBQRModelConfig', 'GBQRRunConfig', 'PoolingStrategy',
+           'PowerTransform', 'SARIXFourierModel', 'SARIXFourierModelConfig', 'SARIXModel', 'SARIXModelConfig',
+           'SARIXRunConfig']
+
+__version__ = '1.3.0'